+++ lib/ui/page/call/widget/participant/widget.dart
@@ -19,13 +19,13 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:medea_jason/medea_jason.dart';
 
-import '../../controller.dart';
 import '../call_cover.dart';
 import '../raised_hand.dart';
 import '../video_view.dart';
 import '/config.dart';
 import '/domain/model/ongoing_call.dart';
 import '/themes.dart';
+import '/ui/page/call/controller.dart';
 import '/ui/page/call/widget/double_bounce_indicator.dart';
 import '/ui/widget/animated_switcher.dart';
 import '/ui/widget/progress_indicator.dart';