+++ CHANGELOG.md
@@ -16,6 +16,9 @@ All user visible changes to this project will be documented in this file. This p
     - Support chat. ([#1595])
     - Chat page:
         - Logs button in notes and support chats. ([#1595])
+- Mobile:
+    - Media panel:
+        - Output device modal being displayed when any external device is connected. ([#1596], [#1593])
 
 ### Fixed
 
@@ -29,6 +32,7 @@ All user visible changes to this project will be documented in this file. This p
 [#1593]: /../../pull/1593
 [#1594]: /../../pull/1594
 [#1595]: /../../pull/1595
+[#1596]: /../../pull/1596